+++ src/lib.rs
@@ -369,9 +369,9 @@ pub mod prelude {
     pub use crate::drawing::*;
     pub use crate::series::{Histogram, LineSeries, PointSeries};
     pub use crate::style::{
-        Black, Blue, Color, Cyan, FontDesc, Green, HSLColor, IntoFont, Magenta, Mixable, Palette,
-        Palette100, Palette99, Palette9999, PaletteColor, RGBColor, Red, ShapeStyle, SimpleColor,
-        TextStyle, Transparent, White, Yellow,
+        Black, Blue, Color, Cyan, FontDesc, FontTransform, Green, HSLColor, IntoFont, Magenta,
+        Mixable, Palette, Palette100, Palette99, Palette9999, PaletteColor, RGBColor, Red,
+        ShapeStyle, SimpleColor, TextStyle, Transparent, White, Yellow,
     };
 
     pub use crate::element::{