+++ src/style/font/mod.rs
@@ -32,6 +32,7 @@ pub trait FontData: Clone {
         _pos: (i32, i32),
         _size: f64,
         _text: &str,
+        _trans: FontTransform,
         _draw: DrawFunc,
     ) -> Result<Result<(), E>, Self::ErrorType> {
         panic!("The font implementation is unable to rasterize font");