+++ src/style/font/ttf.rs
@@ -12,7 +12,7 @@ use lazy_static::lazy_static;
 
 use font_loader::system_fonts;
 
-use super::{FontData, LayoutBox};
+use super::{FontData, FontTransform, LayoutBox};
 
 type FontResult<T> = Result<T, FontError>;
 
@@ -131,15 +131,14 @@ impl FontData for FontDataInternal {
 
         let font = self.0;
 
-        font.layout(text, scale, point(0 as f32, 0 as f32))
-            .for_each(|g| {
-                if let Some(rect) = g.pixel_bounding_box() {
-                    min_x = min_x.min(rect.min.x);
-                    min_y = min_y.min(rect.min.y);
-                    max_x = max_x.max(rect.max.x);
-                    max_y = max_y.max(rect.max.y);
-                }
-            });
+        font.layout(text, scale, point(0.0, 0.0)).for_each(|g| {
+            if let Some(rect) = g.pixel_bounding_box() {
+                min_x = min_x.min(rect.min.x);
+                min_y = min_y.min(rect.min.y);
+                max_x = max_x.max(rect.max.x);
+                max_y = max_y.max(rect.max.y);
+            }
+        });
 
         if min_x == i32::MAX || min_y == i32::MAX {
             return Ok(((0, 0), (0, 0)));
@@ -152,20 +151,27 @@ impl FontData for FontDataInternal {
         (x, y): (i32, i32),
         size: f64,
         text: &str,
+        trans: FontTransform,
         mut draw: DrawFunc,
     ) -> Result<Result<(), E>, Self::ErrorType> {
-        let ((_, b), (_, _)) = self.estimate_layout(size, text)?;
+        //let ((_, b), (_, _)) = self.estimate_layout(size, text)?;
+        let layout = self.estimate_layout(size, text)?;
 
         let scale = Scale::uniform(size as f32);
         let mut result = Ok(());
         let font = self.0;
-        for g in font.layout(text, scale, point(x as f32, y as f32 - b as f32)) {
+
+        let base_x = x + trans.offset(layout).0;
+        let base_y = y + trans.offset(layout).1;
+
+        for g in font.layout(text, scale, point(0.0, 0.0)) {
             if let Some(rect) = g.pixel_bounding_box() {
                 let x0 = rect.min.x;
-                let y0 = rect.min.y;
+                let y0 = rect.min.y - (layout.0).1;
                 g.draw(|x, y, v| {
-                    if x as i32 + x0 >= 0 && y as i32 + y0 >= 0 && result.is_ok() {
-                        result = draw(x as i32 + x0, y as i32 + y0, v);
+                    let (x, y) = trans.transform(x as i32 + x0, y as i32 + y0);
+                    if x + base_x >= 0 && y + base_y >= 0 && result.is_ok() {
+                        result = draw(x + base_x, y + base_y, v);
                     }
                 });
             }